Don't use a completion function with {ivy,helm}-mode
We use a completion function to make the built-in completion respect
the collection's order.  This wasn't necessary for completion with
helm-mode or ivy-mode, but it didn't hurt either.  See
aa5f098a (completion: override default alphabetical sorting,
2017-03-30) for more information.

As of helm's 1aca3e84 (Always use a sync handler when COLLECTION is a
function, 2018-05-24), fuzzy matching no longer works when the
collection is a function, so let's not use the completion function
when we know we don't need it.

Fixes #3476.